# [ThunderFX report] Adds `thunderfx_benchmark_report` to do some automatic performance analysis on the FX graph