… thunder/dynamo/repro_script_template.py
… thunder/dynamo/utils.py
… thunder/tests/test_dynamo.py