<details>
  <summary><b>Before submitting</b></summary>

- [ ] Was this discussed/approved via a Github issue? (no need for typos and docs improvements)
- [ ] Did you read the [contributor guideline](https://github.com/Lightning-AI/pytorch-lightning/blob/main/.github/CONTRIBUTING.md), Pull Request section?
- [ ] Did you make sure to update the docs?
- [ ] Did you write any new necessary tests?

</details>

## What does this PR do?

This PR:
1. Adds `thunderfx_benchmark_report` function that uses the previously implemented reporting APIs to analyze the the input callable, see the function comment for details
<details>
<summary>example outputs</summary>

```
The input callable can be successfully executed by torch.compile.

graph0: Split Information:
The original graph is not split, and is entirely run by Thunder.


graph0_thunder_0 can be successfully executed by Thunder

Benchmark walltime for **graph0_thunder_0 forward** requires investigation: thunder(132.469 us) and torchcompile(28.990 us) is not close (rtol=0.1, atol=0.0)
Benchmark walltime for **graph0_thunder_0 backward** requires investigation: thunder(98.921 us) and torchcompile(34.108 us) is not close (rtol=0.1, atol=0.0)
The scripts are saved: graph0_thunder_0_torchcompile_walltime.py, graph0_thunder_0_thunder_walltime.py


Benchmark kerneltime ran successfully on **graph0_thunder_0 forward**: thunder(955.523 ns) , torchcompile(1.002 us)
Benchmark kerneltime ran successfully on **graph0_thunder_0 backward**: thunder(965.957 ns) , torchcompile(1.012 us)


Benchmark walltime for **graph0_thunder_0_nvFusion0_forward** requires investigation: nvfuser(22.932 us) and torchcompile(20.361 us) is not close (rtol=0.1, atol=0.0)
The scripts are saved: graph0_thunder_0_nvFusion0_forward_torchcompile_walltime.py, graph0_thunder_0_nvFusion0_forward_nvfuser_walltime.py


Benchmark kerneltime ran successfully on **graph0_thunder_0_nvFusion0_forward**: nvfuser(952.301 ns) , torchcompile(998.566 ns)


Benchmark walltime for **graph0_thunder_0_nvFusion0_backward** requires investigation: nvfuser(26.004 us) and torchcompile(21.240 us) is not close (rtol=0.1, atol=0.0)
The scripts are saved: graph0_thunder_0_nvFusion0_backward_torchcompile_walltime.py, graph0_thunder_0_nvFusion0_backward_nvfuser_walltime.py


Benchmark kerneltime ran successfully on **graph0_thunder_0_nvFusion0_backward**: nvfuser(962.118 ns) , torchcompile(1.007 us)

```

</details>

2. Adds `TorchInductorSpecification` which uses only TorchInductor without TorchDynamo, the reason is mentioned here https://github.com/Lightning-AI/lightning-thunder/issues/1521
3. Removes the previous write_repro interface, keeping only the one that takes "CompileSpecification" and timer function as inputs
4. Makes some revisions according to the previous review comments
5. Adds the WAR for Triton error:
  <details>
  <summary>repro script</summary>
  
  ```
  import torch
  
  def test_graph0():
      class DynamoModule(torch.nn.Module):
          def forward(self, L_x_ : torch.Tensor):
              l_x_ = L_x_
              x = l_x_.exp();  l_x_ = None
              return (x,)
          
  
      inputs = [
          torch.testing.make_tensor((2, 2), dtype=torch.float32,  device='cuda:0', requires_grad=True),
      ]
  
      model = DynamoModule()
      compiled_model = torch.compile(model, )
      from thunder.dynamo.report import run_forward_backward
  
      result = compiled_model(*inputs)
  
      output_grads = [torch.ones_like(r) for r in result]
  
      torch.autograd.backward(result, output_grads)
      print(result)
  
  # the WAR: https://github.com/pytorch/pytorch/issues/124565
  # torch.empty(1, device='cuda', requires_grad=True).backward()
  test_graph0()
  ```
  
  <\details>
## PR review

Anyone in the community is free to review the PR once the tests have passed.
If we didn't discuss your PR in Github issues there's a high chance it will not be merged.

## Did you have fun?

Make sure you had fun coding 🙃