@mruberry 2025-02-19
> ```
> def report(model, inputs...):
>     try to run torch.compile
>     # loop subgraphs
>     reports = fx_report(model)
>     for sub_report in reports.sub_reports:
>         thunder_report = analyse_thunder_splits(sub_report)
>         for split_report in thunder_report.sub_reports:
>             ###this part creates new random input tensors, it causes OOM, needs to free the original model and inputs to get enough space
>             split_report.run_benchmark(torchcompile)  
>             split_report.run_benchmark(thunder)
>             compare and save the benchmark script if necessary
> ```
> 
> but it's hard to free the model and inputs inside the function. I'm thinking of providing some utility function and let the user to free the model and inputs between the `get_subgraph_reports` and `performance_analyse`, like:
> 
> ```
> def get_subgraph_reports(model, inputs,...):
>     report_list = []
>     reports = fx_report(model)
>     ...
>     return report_list
> 
> # user code:
> model = GPT(config)
> inputs = make_tensor(...)
> reports = get_subgraph_reports(model, inputs)
> del model, input
> performance_analyse(reports)
> ```
> 
> WDYT?

It makes a lot of sense that we want to get rid of the model and the inputs before starting our analysis, since each report can synthetically create the appropriate inputs and be run in isolation. We can also implement this pattern ourselves in the UX we expose to generate the reports.

Why a helper function vs. just documenting that if OOM is encountered the practitioner should try freeing other cuda tensors?