@kiya00 2025-02-21
Hi @mruberry , it's ready to review, I keep the `thunderfx_benchmark_report` for easy use and give comments about if it goes OOM user needs to delete the model and inputs manually
https://github.com/Lightning-AI/lightning-thunder/blob/ea56d4e87645f1c45c96de24e94e806efae0aea3/thunder/dynamo/report.py#L1055-L1076